+++ eui-client-side/sw.js
@@ -1,1 +1,1 @@
-import 'https://unpkg.com/hra-api@0.9.0/dist/sw.js';
+import 'https://unpkg.com/hra-api@0.10.0/dist/sw.js';